\name{copulaedas-package}

\docType{package}

\alias{copulaedas-package}
\alias{copulaedas}

\title{Estimation of Distribution Algorithms based on Copula Theory}

\description{
This package contains implementations of Estimation of Distribution 
Algorithms (EDAs) based on copula theory. EDAs can be implemented by 
defining methods for generic functions that specify its main parts: 
seeding, selection, learning, sampling, replacement, local optimization, 
termination and reporting methods. It also contains implementations of 
continuous EDAs based on copula theory, a group of well-known continuous 
optimization test problems, and utility functions to analyze the results 
of EDAs. The modular design of the package was inspired by the Mateda-2.0 
toolbox for MATLAB.
}

\references{
Larrañaga, P. and Lozano, J. A. (2002). Estimation of Distribution
Algorithms. A new tool for Evolutionary Computation, Kluwer Academic
Publisher.

Lozano, J. A. and Larrañaga, P. and Inza, I. and Bengoetxea, E. (2006).
Towards a new Evolutionary Computation. Advances in the Estimation of
Distribution Algorithms, Springer-Verlag.

Mühlenbein, H. and Paaß, G. (1996). From recombination of genes to the
estimation of distributions I. Binary parameters, \emph{Parallel Problem
Solving from Nature --- PPSN IV}, 1141, 178--187, Springer-Verlag.

Mühlenbein, H. and Bendisch, J. and Voigt, H. (1996). From recombination
of genes to the estimation of distributions II. Continuous parameters,
\emph{Parallel Problem Solving from Nature --- PPSN IV}, 1141,
188--197, Springer-Verlag.

Santana, R. and Bielza, C. and Larrañaga, P. and Lozano, J. A. and
Echegoyen, C. and Mendiburu, A. and Armañanzas, R. and Shakya, S.
(2010). Mateda-2.0: A MATLAB package for the implementation and analysis
of Estimation of Distribution Algorithms. \emph{Journal of Statistical
Software} \bold{35}, 1--30.
}

\keyword{package}
\keyword{optimize}
