\name{edaCriticalPopSize}

\alias{edaCriticalPopSize}

\title{Critical Population Size}

\description{
Determine the critical population size using a bisection method.
}

\usage{
edaCriticalPopSize(eda, f, lower, upper, fEval, fEvalTol,
                   totalRuns = 30, successRuns = totalRuns,
                   lowerPop = 2, upperPop = NA, stopPercent = 10,
                   verbose = FALSE)
}

\arguments{
\item{eda}{
}
\item{f}{
}
\item{lower}{
}
\item{upper}{
}
\item{fEval}{
}
\item{fEvalTol}{
}
\item{totalRuns}{
}
\item{successRuns}{
}
\item{lowerPop}{
}
\item{upperPop}{
}
\item{stopPercent}{
}
\item{verbose}{
}
}

\details{
}

\value{
}

\references{
Pelikan, M. (2005). Hierarchical Bayesian Optimization Algorithm. 
Toward a new generation of Evolutionary Algorithms, Springer-Verlag.
}

\seealso{
\code{\linkS4class{EDA}},
\code{\link{edaRun}}.
}

\examples{
}
