\name{edaIndepRuns}

\alias{edaIndepRuns}

\title{Independent Runs}

\description{}

\usage{
edaIndepRuns(eda, f, lower, upper, runs, verbose = FALSE)
}

\arguments{
\item{eda}{
}
\item{f}{
}
\item{lower}{
}
\item{upper}{
}
\item{runs}{
}
\item{verbose}{
}
}

\details{
}

\value{
}

\seealso{
\code{\linkS4class{EDA}},
\code{\link{edaRun}}.
}

\examples{
}
