\name{edaOptimize}

\alias{edaOptimize}
\alias{edaOptimizeDisabled}

\title{Local Optimization Methods}

\description{
}

\usage{
edaOptimizeDisabled(eda, gen, pop, popEval, f, lower, upper)
}

\arguments{
\item{eda}{
}
\item{gen}{
}
\item{pop}{
}
\item{popEval}{
}
\item{f}{
}
\item{lower}{
}
\item{upper}{
}
}

\details{
}

\value{
\item{pop}{
}
\item{popEval}{
}
}

\examples{
}
