\name{edaReport}

\alias{edaReport}
\alias{edaReportDisabled}
\alias{edaReportSimple}
\alias{edaReportCombined}

\title{Reporting Methods}

\description{
}

\usage{
edaReportDisabled(eda, gen, fEvals, model, pop, popEval)
edaReportSimple(eda, gen, fEvals, model, pop, popEval)
edaReportCombined(...)
}

\arguments{
\item{eda}{
}
\item{gen}{
}
\item{fEvals}{
}
\item{model}{
}
\item{pop}{
}
\item{popEval}{
}
\item{\dots}{
}
}

\details{
}

\value{
}

\examples{
}
