\name{EDAResults-class}

\docType{class}

\alias{EDAResults-class}
\alias{show,EDAResults-method}
\alias{summary,EDAResults-method}

\title{Class for the Results of Various Runs of an EDA}

\description{
}

\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("EDAResults", ...)}.
}

\section{Methods}{
\describe{
  \item{show}{\code{signature(object = "EDAResults")}: ... }
  \item{summary}{\code{signature(object = "EDAResults")}: ... }
}
}

\seealso{
\code{\linkS4class{EDA}},
\code{\link{edaIndepRuns}}.
}

\keyword{classes}
