\name{margin-norm}

\alias{fnorm}

\title{Normal Marginal Distribution}

\description{
}

\usage{
fnorm(x, lower, upper)
}

\arguments{
\item{x}{
}
\item{lower}{
}
\item{upper}{
}
}

\details{
}

\value{
\item{mean}{
}
\item{sd}{
}
}

\seealso{
\code{\link{pnorm}},
\code{\link{qnorm}}.
}

\examples{
}
