\name{copulaedas-package}

\docType{package}

\alias{copulaedas-package}
\alias{copulaedas}

\title{Estimation of Distribution Algorithms based on Copula Theory}

\description{
This package contains implementations of various classes of Estimation
of Distribution Algorithms (EDAs) based on copula theory: Copula EDAs and Vine
EDAs. In this package, EDAs are implemented using S4 classes with generic
functions for its main parts: seeding, selection, learning, sampling, 
replacement, local optimization, termination, and reporting. This modular 
design was inspired by the Mateda-2.0 toolbox for MATLAB. The package also 
includes the implementation of a group of well-known continuous optimization 
test problems and utility functions to study the behavior of EDAs.
}

\references{
Arderí, R. J. (2007). Algoritmo con Estimación de Distribuciones con Cópula
Gaussiana, Diploma Thesis, University of Havana.

Larrañaga, P. and Lozano, J. A. (2002). Estimation of Distribution
Algorithms. A new tool for Evolutionary Computation, Kluwer Academic
Publisher.

Mühlenbein, H. and Paaß, G. (1996). From recombination of genes to the
estimation of distributions I. Binary parameters, \emph{Parallel Problem
Solving from Nature --- PPSN IV}, 1141, 178--187, Springer-Verlag.

Santana, R. and Bielza, C. and Larrañaga, P. and Lozano, J. A. and
Echegoyen, C. and Mendiburu, A. and Armañanzas, R. and Shakya, S.
(2010). Mateda-2.0: A MATLAB package for the implementation and analysis
of Estimation of Distribution Algorithms. \emph{Journal of Statistical
Software} \bold{35}, 1--30.

Soto, M. and Ochoa, A. and Arderí, R. J. (2007). El Algoritmo con Estimación
de Distribuciones basado en Cópula Gausiana, Technical Report ICIMAF 2007-406,
ISSN 0138-8916, Institute of Cybernetics, Mathematics, and Physics.

Soto, M. and González-Fernández, Y. (2010). Vine Estimation of Distribution
Algorithms, Technical Report ICIMAF 2010-561, ISSN 0138-8916, Institute of 
Cybernetics, Mathematics, and Physics.
}

\keyword{package}
\keyword{optimize}
