\name{edaReplace}

\alias{edaReplace}
\alias{edaReplaceComplete}

\title{Replacement Methods}

\description{
}

\usage{
edaReplaceComplete(eda, gen, pop, popEval, sampledPop, sampledEval)
}

\arguments{
\item{eda}{
}
\item{gen}{
}
\item{pop}{
}
\item{popEval}{
}
\item{sampledPop}{
}
\item{sampledEval}{
}
}

\details{
}

\value{
\item{pop}{
}
\item{popEval}{
}
}

\examples{
}
