\name{edaRun}

\alias{edaRun}
\alias{edaLearn}
\alias{edaSample}

\title{Main Loop of an EDA}

\description{}

\usage{
edaRun(eda, f, lower, upper)
}

\arguments{
\item{eda}{
}
\item{f}{
}
\item{lower}{
}
\item{upper}{
}
}

\details{
}

\value{
}

\seealso{
\code{\linkS4class{EDA}},
\code{\link{edaIndepRuns}}.
}

\examples{
}
