\name{edaTerminate}

\alias{edaTerminate}
\alias{edaTerminateMaxGen}
\alias{edaTerminateMaxEvals}
\alias{edaTerminateEval}
\alias{edaTerminateEvalStdDev}
\alias{edaTerminateCombined}

\title{Termination Methods}

\description{
}

\usage{
edaTerminateMaxGen(eda, gen, fEvals, pop, popEval)
edaTerminateMaxEvals(eda, gen, fEvals, pop, popEval)
edaTerminateEval(eda, gen, fEvals, pop, popEval)
edaTerminateEvalStdDev(eda, gen, fEvals, pop, popEval)
edaTerminateCombined(...)
}

\arguments{
\item{eda}{
}
\item{gen}{
}
\item{fEvals}{
}
\item{pop}{
}
\item{popEval}{
}
\item{\dots}{
}
}

\details{
}

\value{
}

\examples{
}
