\name{margin-empirical}

\alias{fempirical}
\alias{pempirical}
\alias{qempirical}

\title{Kernel-smoothed Empirical Marginal Distribution}

\description{
}

\usage{
fempirical(x, lower, upper)
pempirical(q, X, h)
qempirical(p, X, h)
}

\arguments{
\item{x}{
}
\item{lower}{
}
\item{upper}{
}
\item{q}{
}
\item{p}{
}
\item{X}{
}
\item{h}{
}
}

\details{
}

\value{
}

\examples{
}
